requirements.txt
BabelViscoFDTD>=0.9.10-1
h5py==3.7.0
hdf5plugin==3.3.1
metalcomputebabel @ git+https://github.com/ProteusMRIgHIFU/py-metal-compute.git@90a8112370c571cf8aacf70aa70102e097fb546f
matplotlib==3.6.2
nibabel==4.0.2
numba==0.56.2
numpy-stl==2.17.1
pandas==1.5.0
pydicom==2.3.0
pyinstaller==5.6.2
pyopencl==2022.2.3
PySide6==6.4.0.1
PyYAML==6.0
scikit-image==0.19.3
trimesh==3.12.9
antspyx==0.3.4
superqt==0.4.0
vtk==9.1.0
pyvista==0.37.0
histoprint==2.4.0
linetimer==0.1.5
QtPy==2.3.0

